requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2

# API and Search
google-search-results==2.4.2
google-generativeai==0.3.1

# Video and Audio Processing
ffmpeg-python==0.2.0
openai-whisper==20231117
pyannote.audio==3.1.1
resemblyzer==0.1.1.dev0

# Frontend
streamlit==1.28.2

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
numpy==1.26.2
pandas==2.1.3
gunicorn==21.2.0 